Use PORT value from .env file
Set up a .env file to hold the PORT value. The server will use this
value if it exists and no other PORT envvar is already set in the
current environment. If there is no PORT envvar in the current
environment or in the .env file, the default value of 9000 is used.